requirements.txt
openai
python-dotenv
rich
PyPDF2
flask
flask-cors
python-multipart 